src/main/java/ru/job4j/cinema/repository/Sql2oFileRepository.java
package ru.job4j.cinema.repository;

import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.stereotype.Repository;
import org.sql2o.Sql2o;
import ru.job4j.cinema.model.File;

import java.util.Optional;

@Repository
public class Sql2oFileRepository implements FileRepository {

    private final Sql2o sql2o;

    private static final Logger LOG = LoggerFactory.getLogger(Sql2oFileRepository.class.getName());


    public Sql2oFileRepository(Sql2o sql2o) {
        this.sql2o = sql2o;
    }

    @Override
    public Optional<File> findById(int id) {
        Optional<File> rsl = Optional.empty();
        try (var connection = sql2o.open()) {
            var sql = """
                    SELECT * FROM files WHERE id = :id;
                    """;
            var query = connection.createQuery(sql);
            query.addParameter("id", id);
            var file = query.executeAndFetchFirst(File.class);
            rsl =  Optional.of(file);
        } catch (Exception e) {
            LOG.error("Файл по указананному ID не найден.", e);
        }
        return rsl;
    }

    @Override
    public File save(File file) {
        try (var connection = sql2o.open()) {
            var query = connection.createQuery(
                    "INSERT INTO files (name, path) VALUES (:name, :path)", true)
                    .addParameter("name", file.getName())
                    .addParameter("path", file.getPath());
            int generatedId = query.executeUpdate().getKey(Integer.class);
            file.setId(generatedId);
            return file;
        }
    }

    @Override
    public void deleteById(int id) {
        try (var connection = sql2o.open()) {
            var query = connection.createQuery(
                    "DELETE FROM files WHERE id = :id");
            query.addParameter("id", id).executeUpdate();
        }
    }
}